+++ changelog.md
@@ -1,3 +1,6 @@
 ## Changelog for 2.7.0 (1.21.1+)
 
 - Parity with latest 26.2 features & translations
+  - Bridging Outline is now ON by default. 
+- Expanded Compatibility API
+  - Sable Support